backend/restau_4227/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==36.0.2
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.15
django-allauth==0.51.0
django-ckeditor==6.7.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.4.1
djangorestframework==3.11.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==20.0.4
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
ipython==8.25.0
parso==0.8.4
phonenumbers==8.13.37
pillow==10.3.0
psycopg2==2.8.6
redis==3.3.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
